refactor double to float

+++ LuxTracer/LuxTracer/include/Camera.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <Math/dvec3.h>
+#include <Math/vec3.h>
 #include <Ray.h>
 
 namespace lux
@@ -9,24 +9,24 @@ namespace lux
 	{
 		Camera()
 		{
-			double aspect = 16.0 / 9.0;
-			double ndcHeight = 2.0;
-			double ndcWidth = aspect * ndcHeight;
-			double focalLength = 1.0;
+			float aspect = 16.0f / 9.0f;
+			float ndcHeight = 2.0f;
+			float ndcWidth = aspect * ndcHeight;
+			float focalLength = 1.0f;
 
-			horizontal = math::dvec3(ndcWidth, 0.0, 0.0);
-			vertical = math::dvec3(0.0, ndcHeight, 0.0);
-			btmLeftNDC = position - horizontal * 0.5 - vertical * 0.5 - math::dvec3(0.0, 0.0, focalLength);
+			horizontal = math::vec3(ndcWidth, 0.0f, 0.0f);
+			vertical = math::vec3(0.0f, ndcHeight, 0.0f);
+			btmLeftNDC = position - horizontal * 0.5f - vertical * 0.5f - math::vec3(0.0f, 0.0f, focalLength);
 		}
 
-		Ray const GetRay(double u, double v) const
+		Ray const GetRay(float u, float v) const
 		{
 			return Ray(position, btmLeftNDC + u * horizontal + v * vertical - position);
 		}
 	
-		math::dvec3 position{};
-		math::dvec3 btmLeftNDC{};
-		math::dvec3 horizontal{};
-		math::dvec3 vertical{};
+		math::vec3 position{};
+		math::vec3 btmLeftNDC{};
+		math::vec3 horizontal{};
+		math::vec3 vertical{};
 	};
 }
